pyproject.toml
[build-system]
requires = ["scikit_build_core[pyproject]", "pybind11"]
build-backend = "scikit_build_core.build"

[tool.scikit-build]
build-dir = "src/am4/utils/build"
cmake.args = ["-DEXCLUDE_EXECUTABLES=1"]
cmake.build-type = "Release"
cmake.source-dir = "src/am4/utils"
cmake.verbose = true
logging.level = "DEBUG"
wheel.install-dir = "am4"
wheel.expand-macos-universal-tags = true
sdist.exclude = ["misc/demand-research", "src/am4/utils/data/*.parquet"]

[project]
name = "am4"
version = "0.1.10-alpha.1"
description = "Tools and utilities for the game Airline Manager 4."
authors = [{ name = "cathaypacific8747", email = "abraham@ylcheung.com" }]
license = { text = "MIT License" }
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.9",
]
dependencies = [
    "rich>=13.7.0",
    "httpx>=0.26.0",
    "pydantic>=2.6.1",
    "loguru>=0.7.2",
    "typer>=0.9.0",
    "orjson>=3.9.13",
]

[project.optional-dependencies]
dev = [
    # cpp
    "pybind11>=2.11.1",
    "pybind11-stubgen>=2.4.2",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.4",
    # py
    "ruff",
]
docs = [
    "mkdocs-material[imaging]",
    "mkdocstrings[python]",
    "black",
    "mkdocs-exclude",
]
api = ["fastapi>=0.109.2", "uvicorn>=0.27.0"]
bot = [
    "matplotlib>=3.9.1",
    "Pillow>=10.3.0",
    "discord.py>=2.4.0",
    "pyproj>=3.6.1",
    "cmocean>=3.1.3",
    "pyarrow>=16.1.0",
]
research = ["pandas", "PyQt6", "tqdm", "ipykernel", "scipy", "polars"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals"]
testpaths = ["src/am4/utils/tests"]


[tool.ruff]
select = ["E", "F"]
ignore = [
    "PT011",
    "W292",
    "E741",  # ambiguous variable name
]
exclude = ["src-old", "src-v2-v3"]
line-length = 120
target-version = "py39"

[tool.ruff.per-file-ignores]
"tests/**" = ["T20"]

[tool.cibuildwheel]
test-requires = "pytest"
test-command = "pytest"
build-verbosity = 1
